requirements.txt
alembic==1.2.1
SQLAlchemy==1.3.8
PyMySQL==0.9.3
python-dotenv
graphene
python-i18n[YAML]==0.3.1
sqlalchemy-utils
graphene-sqlalchemy>=2.0
requests
boto3
SQLAlchemy-serializer
xmltodict
factory_boy==2.9.2
pytz